modularizamos el código y añadimos un makefile para compilar todo a la vez y limpiar los *.o restantes de la compilación
